Just Confirming that everything is updated

+++ lab4/ErrorCheckListCode/src/ErrorCheckList/BankAccount.java
@@ -12,7 +12,7 @@ public class BankAccount {
          amount = startingAmount; 
     }
 
-    public void deposit(double depositAmount) {
+    public void deposit(int depositAmount) {
         balance += depositAmount; 
     }
